Add public missing keys method to formatter
When a named route cannot be resolved due to a missing key, the router should be capable of identifying the required keys for the named route. This will be extremely useful in debugging routing errors by providing enough information to fix the problem in the message.

Functionality is tested, ATP